requirements.txt
torch
torchvision
streamlit
numpy
Pillow
matplotlib
